+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## TBD
 
+* Delete cached error reports if an Exception is thrown during disk IO, preventing delivery of empty/partial reports on the next app launch.
+  [#609](https://github.com/bugsnag/bugsnag-android/pull/609)
+
 * Prevent internal error reporting of FileNotFoundException during serialization
   [#605](https://github.com/bugsnag/bugsnag-android/pull/605)